+++ source/D2Common/include/D2Collision.h
@@ -12,6 +12,27 @@ struct D2BoundingBoxStrc
 	int32_t nTop;								//0x0C
 };
 
+enum D2C_CollisionFlags
+{
+	COLLIDE_NONE = 0x0000,
+	COLLIDE_BLOCK_PLAYER = 0x0001,			// 'black space' in arcane sanctuary, cliff walls etc
+	COLLIDE_BLOCK_MISSILE = 0x0002,			// tile based obstacles you can't shoot over
+	COLLIDE_WALL = 0x0004,					// again used inconsistantly -.-
+	COLLIDE_BLOCK_LEAP = 0x0008,
+	COLLIDE_ALTERNATE_FLOOR = 0x0010,		// some floors have this set, others don't
+	COLLIDE_BLANK = 0x0020,					// returned if the subtile is invalid
+	COLLIDE_MISSILE = 0x0040,
+	COLLIDE_PLAYER = 0x0080,
+	COLLIDE_MONSTER = 0x0100,
+	COLLIDE_ITEM = 0x0200,
+	COLLIDE_OBJECT = 0x0400,
+	COLLIDE_DOOR = 0x0800,
+	COLLIDE_UNIT_RELATED = 0x1000,			// set for units sometimes, but not always
+	COLLIDE_PET = 0x2000,
+	COLLIDE_4000 = 0x4000,
+	COLLIDE_CORPSE = 0x8000,				// also used by portals, but dead monsters are mask 0x8000
+};
+
 //D2Common.0x6FD41000
 void __fastcall D2Common_COLLISION_FirstFn_6FD41000(D2RoomStrc* pRoom, D2DrlgTileDataStrc* pTileData, D2TileLibraryEntryStrc* pTileLibraryEntry);
 //D2Common.0x6FD411F0
