+++ source/D2Common/include/D2Composit.h
@@ -2,6 +2,44 @@
 
 #include "CommonDefinitions.h"
 
+enum D2Composits
+{
+	COMPOSIT_HEAD,
+	COMPOSIT_TORSO,
+	COMPOSIT_LEGS,
+	COMPOSIT_RIGHTARM,
+	COMPOSIT_LEFTARM,
+	COMPOSIT_RIGHTHAND,
+	COMPOSIT_LEFTHAND,
+	COMPOSIT_SHIELD,
+	COMPOSIT_SPECIAL1,
+	COMPOSIT_SPECIAL2,
+	COMPOSIT_SPECIAL3,
+	COMPOSIT_SPECIAL4,
+	COMPOSIT_SPECIAL5,
+	COMPOSIT_SPECIAL6,
+	COMPOSIT_SPECIAL7,
+	COMPOSIT_SPECIAL8,
+	NUM_COMPONENTS,
+};
+
+enum D2WeaponClasses
+{
+	WEAPONCLASS_NONE,
+	WEAPONCLASS_BOW,
+	WEAPONCLASS_1HS,
+	WEAPONCLASS_1HT,
+	WEAPONCLASS_STF,
+	WEAPONCLASS_2HS,
+	WEAPONCLASS_2HT,
+	WEAPONCLASS_XBW,
+	WEAPONCLASS_1JS,
+	WEAPONCLASS_1JT,
+	WEAPONCLASS_1SS,
+	WEAPONCLASS_1ST,
+	WEAPONCLASS_HT1,
+	WEAPONCLASS_HT2
+};
 
 //D2Common.0x6FD466C0 (#10884)
 void __stdcall D2Common_10884_COMPOSIT_unk(D2UnitStrc* pUnit, int nClass, int nMode, int nUnitType, D2InventoryStrc* pInventory, char* szPath, int* pWeaponClassCode, BOOL bAddPathPrefix, int a9);
