+++ source/D2Common/include/DataTbls/ObjectsTbls.h
@@ -2,6 +2,17 @@
 
 #include <D2BasicTypes.h>
 
+enum D2ObjectSubClasses
+{
+	OBJSUBCLASS_SHRINE = 0x01,
+	OBJSUBCLASS_OBELISK = 0x02,
+	OBJSUBCLASS_TOWNPORTAL = 0x04,
+	OBJSUBCLASS_CHEST = 0x08,
+	OBJSUBCLASS_PORTAL = 0x10,
+	OBJSUBCLASS_WELL = 0x20,
+	OBJSUBCLASS_WAYPOINT = 0x40,
+	OBJSUBCLASS_DOOR = 0x80,
+};
 
 struct D2ObjectsTxt
 {
