+++ source/D2Common/include/Drlg/D2DrlgDrlg.h
@@ -29,6 +29,50 @@ enum D2AltDirections
 	ALTDIR_SOUTHWEST,
 };
 
+enum D2C_DrlgTypes
+{
+	DRLGTYPE_MAZE = 0x01,
+	DRLGTYPE_PRESET = 0x02,
+	DRLGTYPE_OUTDOOR = 0x03,
+	NUM_DRLGTYPES
+};
+
+enum D2RoomExFlags
+{
+	ROOMEXFLAG_INACTIVE = 0x00000002,
+	ROOMEXFLAG_HAS_WARP_0 = 0x00000010,
+	ROOMEXFLAG_HAS_WARP_1 = 0x00000020,
+	ROOMEXFLAG_HAS_WARP_2 = 0x00000040,
+	ROOMEXFLAG_HAS_WARP_3 = 0x00000080,
+	ROOMEXFLAG_HAS_WARP_4 = 0x00000100,
+	ROOMEXFLAG_HAS_WARP_5 = 0x00000200,
+	ROOMEXFLAG_HAS_WARP_6 = 0x00000400,
+	ROOMEXFLAG_HAS_WARP_7 = 0x00000800,
+	ROOMEXFLAG_SUBSHRINE_ROW1 = 0x00001000,
+	ROOMEXFLAG_SUBSHRINE_ROW2 = 0x00002000,
+	ROOMEXFLAG_SUBSHRINE_ROW3 = 0x00004000,
+	ROOMEXFLAG_SUBSHRINE_ROW4 = 0x00008000,
+	ROOMEXFLAG_HAS_WAYPOINT = 0x00010000,			//outdoors with subtheme and subwaypoint
+	ROOMEXFLAG_HAS_WAYPOINT_SMALL = 0x00020000,		//waypoint small
+	ROOMEXFLAG_AUTOMAP_REVEAL = 0x00040000,
+	ROOMEXFLAG_NO_LOS_DRAW = 0x00080000,
+	ROOMEXFLAG_HAS_ROOM = 0x00100000,				//an active pRoom structure is attached to this RoomEx
+	ROOMEXFLAG_ROOM_FREED_SRV = 0x00200000,			//set after freeing the pRoom
+	ROOMEXFLAG_HASPORTAL = 0x00400000,				//prevents room deletion
+	ROOMEXFLAG_POPULATION_ZERO = 0x00800000,		//set for towns, also set if a ds1 has populate=0 // nospawn region
+	ROOMEXFLAG_TILELIB_LOADED = 0x01000000,
+	ROOMEXFLAG_PRESET_UNITS_ADDED = 0x02000000,		//refers to DRLGMap or DRLGFile; adds the hardcoded preset units
+	ROOMEXFLAG_PRESET_UNITS_SPAWNED = 0x04000000,	//set after RoomEx preset units have been spawned / prevents respawning them on room reactivation
+	ROOMEXFLAG_ANIMATED_FLOOR = 0x08000000,			//animated floors (river of flame, hell act5)
+
+	ROOMEXFLAG_HAS_WARP = (ROOMEXFLAG_HAS_WARP_0 | ROOMEXFLAG_HAS_WARP_1 | ROOMEXFLAG_HAS_WARP_2 | ROOMEXFLAG_HAS_WARP_3 | ROOMEXFLAG_HAS_WARP_4 | ROOMEXFLAG_HAS_WARP_5 | ROOMEXFLAG_HAS_WARP_6 | ROOMEXFLAG_HAS_WARP_7),
+};
+
+enum D2DrlgLevelFlags
+{
+	DRLGLEVELFLAG_AUTOMAP_REVEAL = 0x10,
+};
+
 //D2Common.0x6FD74120 (#10014)
 D2DrlgStrc* __fastcall DRLG_AllocDrlg(D2DrlgActStrc* pAct, uint8_t nActNo, void* a3, uint32_t nInitSeed, int nLevelId, uint32_t nFlags, D2GameStrc* pGame, uint8_t nDifficulty, AUTOMAPFN pfAutoMap, TOWNAUTOMAPFN pfTownAutoMap);
 //D2Common.0x6FD743B0 (#10012)
