+++ source/D2Common/include/Units/Player.h
@@ -1,0 +1,49 @@
+#pragma once 
+
+
+enum D2C_PlayerModes
+{
+	PLRMODE_DEATH,		// DT
+	PLRMODE_NEUTRAL,	// NU
+	PLRMODE_WALK,		// WL
+	PLRMODE_RUN,		// RN
+	PLRMODE_GETHIT,		// GH
+	PLRMODE_TNEUTRAL,	// TN
+	PLRMODE_TWALK,		// TW
+	PLRMODE_ATTACK1,	// A1
+	PLRMODE_ATTACK2,	// A2
+	PLRMODE_BLOCK,		// BL
+	PLRMODE_CAST,		// SC
+	PLRMODE_THROW,		// TH
+	PLRMODE_KICK,		// KK
+	PLRMODE_SPECIAL1,	// S1
+	PLRMODE_SPECIAL2,	// S2
+	PLRMODE_SPECIAL3,	// S3
+	PLRMODE_SPECIAL4,	// S4
+	PLRMODE_DEAD,		// DD
+	PLRMODE_SEQUENCE,	// SQ
+	PLRMODE_KNOCKBACK	// KB
+};
+
+enum D2C_PlayerClasses
+{
+	PCLASS_AMAZON,
+	PCLASS_SORCERESS,
+	PCLASS_NECROMANCER,
+	PCLASS_PALADIN,
+	PCLASS_BARBARIAN,
+	PCLASS_DRUID,
+	PCLASS_ASSASSIN,
+	PCLASS_EVILFORCE,
+	PCLASS_INVALID = 7,
+	NUMBER_OF_PLAYERCLASSES = 7,
+};
+
+enum D2C_PlayerFlags
+{
+	PLRFLAG_NEWBIE = 0x01,
+	PLRFLAG_HARDCORE = 0x04,
+	PLRFLAG_DEAD = 0x08,
+	PLRFLAG_EXPANSION = 0x20,
+	PLRFLAG_LADDER = 0x40,
+};
