+++ source/D2Common/src/Drlg/DrlgDrlgAnim.cpp
@@ -1,6 +1,7 @@
 #include "Drlg/D2DrlgDrlgAnim.h"
 
 #include "D2DataTbls.h"
+#include "Drlg/D2DrlgDrlg.h"
 #include "Drlg/D2DrlgDrlgGrid.h"
 #include "Drlg/D2DrlgRoomTile.h"
 #include <D2CMP.h>
